Idea Labs fix and merge
