+++ cypress/e2e/ideyalabs/orders/receiving-and-check-in/encumbrance-releases-with-payment-not-required-status.cy.js
@@ -1,0 +1,44 @@
+import orderFragment from '../../../../support/fragments/orders/orders';
+import orderLines from '../../../../support/fragments/orders/orderLines';
+import receiving from '../../../../support/fragments/receiving/receiving';
+import topMenu from '../../../../support/fragments/topMenu';
+import testTypes from '../../../../support/dictionary/testTypes';
+import devTeams from '../../../../support/dictionary/devTeams';
+
+const orderDetails = {
+  searchByParameter: 'PO line number',
+  enterPoLineNumber: '20692-1',
+  checkOrderLineSearchResults: '20692-1',
+  titleName: 'Approve rolled',
+  caption: 'Done',
+  enumeration: 'Electronic',
+  poLineNumber: '20692-1',
+};
+
+
+
+
+describe('Orders: Receiving and Check-in ', () => {
+  before(() => {
+    cy.login(Cypress.env('diku_login'), Cypress.env('diku_password'));
+  });
+
+  it('C378899 Encumbrance releases when receive piece for order with payment status "Payment Not Required" (Thunderjet)', { tags: [testTypes.ideaLabsTests, devTeams.thunderjet] }, () => {
+    cy.visit(topMenu.ordersPath);
+    orderLines.clickOnOrderLines();
+    orderLines.searchByParameter(
+      orderDetails.searchByParameter,
+      orderDetails.enterPoLineNumber
+    );
+    orderLines.checkOrderlineSearchResults(
+      orderDetails.checkOrderLineSearchResults
+    );
+    orderLines.selectOrderline(orderDetails.checkOrderLineSearchResults);
+    orderLines.receiveOrderLineViaActions();
+    orderLines.selectreceivedTitleName(orderDetails.titleName);
+    receiving.addPieceProcess(orderDetails.caption, orderDetails.enumeration);
+    receiving.quickReceivePieceAdd(orderDetails.enumeration);
+    receiving.clickOnPOLnumber(orderDetails.poLineNumber);
+    orderFragment.selectFundIDFromthelist();
+  });
+});
